Move inference callback tests to shared location
Test coverage for build_inference_callback (epoch filtering, weighted
error aggregation, epoch-set skipping, missing-metric errors) was
duplicated between ace and coupled. Consolidate into a single
TestBuildInferenceCallback in fme/core/generics/test_trainer.py.
