+++ fme/coupled/train/test_train_config.py
@@ -434,110 +434,3 @@ class TestGetValidationCallback:
             ],
         )
         assert loss == pytest.approx(2.0 * 0.1 + 3.0 * 0.2)
-
-
-class TestGetInferenceCallback:
-    @staticmethod
-    def _make_entry(name, weight=1.0, n_coupled_steps=1):
-        config = MagicMock()
-        config.weight = weight
-        config.n_coupled_steps = n_coupled_steps
-        data = MagicMock()
-        # data.loader is iterated once to get a batch for initial_times
-        data.loader = iter([MagicMock()])
-        return (config, data, name)
-
-    @staticmethod
-    def _call(
-        entries,
-        inference_one_epoch_side_effect,
-        epoch=1,
-        inference_epochs=(1,),
-        inference_epoch_sets=None,
-    ):
-        from fme.coupled.train.train import get_inference_callback
-
-        if inference_epoch_sets is None:
-            inference_epoch_sets = [{1} for _ in entries]
-        stepper = MagicMock()
-        with patch(
-            "fme.core.generics.trainer.inference_one_epoch",
-            side_effect=inference_one_epoch_side_effect,
-        ):
-            callback = get_inference_callback(
-                inference_entries=entries,
-                inference_epochs=list(inference_epochs),
-                inference_epoch_sets=list(inference_epoch_sets),
-                stepper=stepper,
-                dataset_info=MagicMock(),
-                output_dir="/tmp/out",
-                save_per_epoch_diagnostics=False,
-            )
-            return callback(epoch=epoch)
-
-    def test_epoch_not_in_inference_epochs_returns_empty(self):
-        entries = [self._make_entry("inference")]
-        logs, error = self._call(
-            entries,
-            inference_one_epoch_side_effect=[],
-            epoch=2,
-            inference_epochs=(1,),
-        )
-        assert logs == {}
-        assert error is None
-
-    def test_single_entry_weighted_error(self):
-        entries = [self._make_entry("inference", weight=2.0)]
-        logs, error = self._call(
-            entries,
-            [{"inference/time_mean_norm/rmse/channel_mean": 0.4}],
-        )
-        assert error == pytest.approx(2.0 * 0.4)
-        assert "inference/time_mean_norm/rmse/channel_mean" in logs
-
-    def test_weighted_entry_missing_metric_raises(self):
-        entries = [self._make_entry("a", weight=1.0)]
-        with pytest.raises(RuntimeError, match="did not produce expected metric key"):
-            self._call(
-                entries,
-                [{"a/other_metric": 1.0}],
-            )
-
-    def test_all_zero_weight_returns_none_error(self):
-        entries = [self._make_entry("a", weight=0.0)]
-        logs, error = self._call(
-            entries,
-            [{"a/time_mean_norm/rmse/channel_mean": 0.5}],
-        )
-        assert error is None
-        assert "a/time_mean_norm/rmse/channel_mean" in logs
-
-    def test_multiple_weighted_entries(self):
-        entries = [
-            self._make_entry("a", weight=2.0),
-            self._make_entry("b", weight=3.0),
-        ]
-        logs, error = self._call(
-            entries,
-            [
-                {"a/time_mean_norm/rmse/channel_mean": 0.1},
-                {"b/time_mean_norm/rmse/channel_mean": 0.2},
-            ],
-        )
-        assert error == pytest.approx(2.0 * 0.1 + 3.0 * 0.2)
-
-    def test_entry_skipped_when_not_in_epoch_set(self):
-        entries = [
-            self._make_entry("a", weight=1.0),
-            self._make_entry("b", weight=1.0),
-        ]
-        logs, error = self._call(
-            entries,
-            [{"a/time_mean_norm/rmse/channel_mean": 0.5}],
-            epoch=1,
-            inference_epochs=(1,),
-            inference_epoch_sets=[{1}, {2}],
-        )
-        assert error == pytest.approx(0.5)
-        assert "a/time_mean_norm/rmse/channel_mean" in logs
-        assert "b/time_mean_norm/rmse/channel_mean" not in logs
